+++ src/layouts/DarsSideBar/DarsSideBar.tsx
@@ -1,0 +1,5 @@
+const DarsSideBar = () => {
+  return <div>Course Details</div>;
+};
+
+export default DarsSideBar;
