+++ src/layouts/LeftPanel/LeftPanel.tsx
@@ -1,0 +1,5 @@
+const LeftPanel = () => {
+  return <div>Left Panel</div>;
+};
+
+export default LeftPanel;
