+++ src/types/Course.ts
@@ -1,0 +1,8 @@
+export interface Course {
+  id: number;
+  code: string;
+  name: string;
+  description: string;
+  units: number;
+  instructor: string;
+}
